backend/requirements.txt
# ============================================
# DCAI Intelligence Platform - Backend Dependencies
# ============================================

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0

# Async Support
httpx>=0.24,<0.28
aiofiles>=23.2.1

# Database
supabase>=2.3.0
asyncpg>=0.29.0
sqlalchemy>=2.0.25

# LLM - using httpx directly via our LLMService, not langchain
openai>=1.12.0

# RAGFlow Client - using httpx

# Web Scraping & Search
firecrawl-py>=0.0.16
tavily-python>=0.3.0

# Data Processing
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
orjson>=3.9.0

# File Handling
aiofiles>=23.2.1

# Utilities
tenacity>=8.2.3
structlog>=24.1.0
python-dateutil>=2.8.2
uuid6>=2024.1.12

# Validation
email-validator>=2.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.4
pytest-cov>=4.1.0

# Development
ruff>=0.1.14
